OS timer accuracy changed from 10ms to 1ms (#419)
1. OS timer accuracy changed from 10ms to 1ms

2. improve "printer unconnect status reminder"

+++ TFT/src/User/Menu/Selectmode.h
@@ -14,10 +14,10 @@ typedef enum
   MKEY_IDLE = IDLE_TOUCH,
 }MKEY_VALUES;
 
-bool LCD_ReadPen(uint8_t intervals);
+bool LCD_ReadPen(uint16_t intervals);
 uint8_t LCD_ReadTouch(void);
 void Touch_Sw(uint8_t num);
-bool LCD_BtnTouch(uint8_t intervals);
+bool LCD_BtnTouch(uint16_t intervals);
 
 //select mode fun
 void show_selectICON(void);
